dev: refactor song charts to use common abstract interface
